requirements.txt
customtkinter==5.2.2
Pillow==10.3.0
pytesseract==0.3.10
